requirements.txt
Flask
Flask-CORS
requests
xmltodict
python-dotenv
openai
gunicorn
pymysql